change dec to unsigned long

+++ 0x19-hash_tables/2-key_index.c
@@ -8,8 +8,7 @@
  */
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int c;
-
+	unsigned long int c;
 	c = hash_djb2(key) % size;
 	return (c);
 }
